refactor: Replace WrappedDialog with new Dialog and WarningDialog components
