+++ src/components/ui/dialog/index.ts
@@ -1,0 +1,2 @@
+export * from "./Dialog";
+export * from "./WarningDialog";
