+++ src/components/ui/index.ts
@@ -4,4 +4,4 @@ export * from "./link";
 export * from "./non-upper-button";
 export * from "./progress-bar";
 export * from "./section-with-title";
-export * from "./wrapped-dialog";
+export * from "./dialog";
