--- src/components/ui/wrapped-dialog/index.ts
@@ -1,1 +1,0 @@
-export * from "./WrappedDialog";
